+++ src/main/java/scs/comp5903/cucumber/builder/JFeatureBuilder.java
@@ -7,12 +7,12 @@ import scs.comp5903.cucumber.execution.JScenarioOutline;
 import scs.comp5903.cucumber.execution.MethodExecution;
 import scs.comp5903.cucumber.model.exception.EasyCucumberException;
 import scs.comp5903.cucumber.model.exception.ErrorCode;
-import scs.comp5903.cucumber.model.feature.JFeatureDetail;
-import scs.comp5903.cucumber.model.feature.JScenarioDetail;
-import scs.comp5903.cucumber.model.feature.JScenarioOutlineDetail;
-import scs.comp5903.cucumber.model.feature.jstep.AbstractJStep;
-import scs.comp5903.cucumber.model.stepdef.JStepDefDetail;
-import scs.comp5903.cucumber.model.stepdef.JStepDefMethodDetail;
+import scs.comp5903.cucumber.model.jfeature.JFeatureDetail;
+import scs.comp5903.cucumber.model.jfeature.JScenarioDetail;
+import scs.comp5903.cucumber.model.jfeature.JScenarioOutlineDetail;
+import scs.comp5903.cucumber.model.jfeature.jstep.AbstractJStep;
+import scs.comp5903.cucumber.model.jstepdef.JStepDefDetail;
+import scs.comp5903.cucumber.model.jstepdef.JStepDefMethodDetail;
 
 import java.util.*;
