+++ src/main/java/scs/comp5903/cucumber/model/jfeature/JScenarioOutlineDetail.java
@@ -1,4 +1,4 @@
-package scs.comp5903.cucumber.model.feature;
+package scs.comp5903.cucumber.model.jfeature;
 
 import java.util.List;
 import java.util.Objects;
