+++ src/main/java/scs/comp5903/cucumber/model/jstepdef/matcher/AbstractJStepMatcher.java
@@ -1,4 +1,4 @@
-package scs.comp5903.cucumber.model.stepdef.matcher;
+package scs.comp5903.cucumber.model.jstepdef.matcher;
 
 
 import java.util.Objects;
