+++ src/test/java/scs/comp5903/cucumber/model/jfeature/jstep/AbstractJStepTest.java
@@ -1,4 +1,4 @@
-package scs.comp5903.cucumber.model.feature.jstep;
+package scs.comp5903.cucumber.model.jfeature.jstep;
 
 import org.junit.jupiter.api.Test;
