+++ src/test/java/scs/comp5903/cucumber/parser/JStepDefinitionParserTest.java
@@ -4,9 +4,9 @@ import org.junit.jupiter.api.Test;
 import scs.comp5903.cucumber.model.annotation.step.JStep;
 import scs.comp5903.cucumber.model.annotation.step.JStepKeyword;
 import scs.comp5903.cucumber.model.exception.EasyCucumberException;
-import scs.comp5903.cucumber.model.stepdef.matcher.GivenJStepMatcher;
-import scs.comp5903.cucumber.model.stepdef.matcher.ThenJStepMatcher;
-import scs.comp5903.cucumber.model.stepdef.matcher.WhenJStepMatcher;
+import scs.comp5903.cucumber.model.jstepdef.matcher.GivenJStepMatcher;
+import scs.comp5903.cucumber.model.jstepdef.matcher.ThenJStepMatcher;
+import scs.comp5903.cucumber.model.jstepdef.matcher.WhenJStepMatcher;
 import scs.comp5903.cucumber.parser.samplestepdef.AnotherSampleStepDefinition;
 import scs.comp5903.cucumber.parser.samplestepdef.SampleStepDefWithNewAnnotation;
 import scs.comp5903.cucumber.parser.samplestepdef.SampleStepDefinition;
